+++ src/main/java/com/example/gimmegonghakauth/user/infrastructure/UserRepository.java
@@ -1,12 +1,12 @@
-package com.example.gimmegonghakauth.dao;
+package com.example.gimmegonghakauth.user.infrastructure;
 
-import com.example.gimmegonghakauth.domain.UserDomain;
+import com.example.gimmegonghakauth.user.domain.UserDomain;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserDao extends JpaRepository<UserDomain,Long> {
+public interface UserRepository extends JpaRepository<UserDomain,Long> {
 
     Optional<UserDomain> findByStudentId(Long studentId);
